refactor: split scrabble.py into scrabble.py and move.py
