+++ python/test/test_config.py
@@ -16,10 +16,9 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 """
 
-import json
 import os
-from pathlib import Path
 import unittest
+from pathlib import Path
 
 from config import config
